refactor: 패키지로 기능별로 묶어서 구별
컨트롤러 패키지에 게임기능과 정답기능을 구별하여 추가하고 테스트용 프린트보기와 주석을 제거하여 가독성을 높였습니다.
