+++ src/main/java/baseball/controller/answer/AnswerController.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.controller.answer;
 
 import camp.nextstep.edu.missionutils.Randoms;
