+++ src/main/java/baseball/controller/game/GameController.java
@@ -1,5 +1,6 @@
-package baseball;
+package baseball.controller.game;
 
+import baseball.controller.answer.AnswerController;
 import baseball.exception.GameException;
 import camp.nextstep.edu.missionutils.Console;
 
@@ -13,7 +14,6 @@ public class GameController {
     public static void gameInitializer() {
         System.out.println("숫자 야구 게임을 시작합니다.");
         canPlay = true;
-
         while (canPlay) {
             theAnswer = AnswerController.makeAnswer();
             System.out.println("answer: " + theAnswer);    // 테스트용
@@ -28,15 +28,8 @@ public class GameController {
     private static void gamePlay() {
         System.out.print("숫자를 입력해주세요 : ");
         guessNum = Console.readLine();
-//        if (guessNum.length() > 3) {
-//            throw new IllegalArgumentException();
-//        }
         GameException.inputExceptionChecker(guessNum);
-        System.out.println("guessNum = " + guessNum);   // 테스트용
-
-        int resultCount = Counter.hitBallCounter(guessNum, theAnswer);
-        System.out.println("resultCount = " + resultCount);     // for test
-
+        int resultCount = GameCounter.hitBallCounter(guessNum, theAnswer);
         playAgain(resultCount);
     }
